[hotfix] findHoldintsByProfile 관련 에러

+++ src/main/java/farmsystem/backend/domain/portfolio/dto/HoldingDto.java
@@ -3,9 +3,9 @@ package farmsystem.backend.domain.portfolio.dto;
 public record HoldingDto(
         String stockCode,
         String stockName,
-        int quantity,
+        long quantity,
         long buyPrice,
-        int buyAmount
+        long buyAmount
 ) {
     public double getAveragePrice() {
         return buyAmount == 0 ? 0 : (double) buyPrice / buyAmount;
